+++ CHANGELOG.md
@@ -5,6 +5,11 @@
 #### 🚀 Updates
 
 - Added colors to command line `--help` menus.
+- Updated the following locations to support partial versions and aliases:
+  - Tool versions in `.prototools`.
+  - Pinning a default version with `proto install --pin`.
+  - Setting global version with `proto global`.
+  - Setting local version with `proto local`.
 - WASM API
   - Added `is_musl` and `get_target_triple` helper functions.